# 설마 long?

+++ src/main/kotlin/cupidcrew/backend/api/dao/crew/CrewEntity.kt
@@ -28,7 +28,7 @@ class CrewEntity(
     val company: String,
 
     @Column(nullable = false)
-    val isApproved: Int,
+    val isApproved: Long,
 ) : BaseTime(), UserDetails {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
