+++ src/main/kotlin/cupidcrew/backend/api/dto/crew/CrewSignupRequestDto.kt
@@ -6,5 +6,5 @@ data class CrewSignupRequestDto(
     var m_password: String? = "",
     val mutualFriend: String,
     val company: String,
-    val isApproved: Int?= 0,
+    val isApproved: Long?= 0,
 )
